+++ CHANGELOG.md
@@ -7,7 +7,10 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [1.0.2]
 
 - 新增变量支持：`${fileName}`、`${extensionPath}`，可在 shell-command 和 npm-script 的 command/script 字段中使用
-- actions.json 增加了一个png文件的OCR示例,OCR 脚本位于插件bin目录下
+
+- 命令语法从command 改为commands,支持多行命令.
+
+- 并且增加了多行示例:OCR
 
 ## [1.0.1]